Refactor metadata fetching scripts and skills
- Deleted the old `kb-meta-fetch` skill and its associated script.
- Introduced new `fetch-meta-crossref2kb` skill with updated functionality for fetching journal articles from Crossref.
- Added `fetch-meta-from-kb` skill to directly fetch recent journal rows from PostgreSQL.
- Implemented new scripts for both skills to handle data fetching and JSON output.
- Updated OpenAI agent configurations for the new skills.
